feat: Add Loading component
This commit adds a new component called Loading, which displays a loading spinner while content is being loaded. The Loading component is used to provide a visual indication to the user that the application is busy fetching data or performing other asynchronous operations. It is positioned fixed at the center of the screen and has a semi-transparent background. The Loading component is implemented using the CircularProgress component from the Material-UI library.

Co-authored-by: piccolojnr <rahimdaud246@gmail.com>