refactor: Remove unused code and update signup process
The code changes in signup.tsx remove unused code related to Redux and login functionality. The dispatch and login functions are no longer needed. Additionally, the signup process is updated to navigate to the login page after successful registration instead of the profile page. This refactor improves code cleanliness and simplifies the signup process.

Co-authored-by: piccolojnr <rahimdaud246@gmail.com>